// src/types/enums.ts
export enum StoryGuid {
  PLAY_GAME = '284a0d3d-af06-482d-942f-8efafbf0d4aa',
  PAUSE_GAME = 'fe3c96e8-432d-4cfd-ad8b-96e30167c99e',
  OBSERVING_INVENTORY = '4cd69f91-45ec-4871-9d96-be8ea01d0c2f',
}

export enum TOOL_NAMES {
  AXE = 'axe',
  // PICKAXE = 'pickaxe',
  // SHOVEL = 'shovel',
  HOE = 'hoe',
}

export enum ITEM_GROUPS {
  TOOLS = 'tools',
  WEAPONS = 'weapons',
  FOOD = 'food',
  CROPS = 'crops',
  CLOTHING = 'clothing',
  POTION = 'potion',
  MATERIALS = 'materials',
  IMPORTANT_ITEMS = 'important_items', 
}

export enum TERRAIN_TILES {
  WATER = 'water',
  GRASS = 'grass',
  DIRT = 'dirt',
  FARMLAND = 'farmland',
}

export enum TileTypes {
  DIRT = 'dirt',
  GRASS = 'grass',
  WATER = 'water',
}